Make stateLogger use same api as actionLogger
